move checks to Matchers file
